feat: upgrade ts-rest, use app router for internal api to allow use of revalidateTage
not sure if this is documented anywhere at all, but you apparently can't use revalidateTag/path in the pages router. since our api is in the pages router, we can't call any of those functions at all.

However, this is necessary, as we need to revalidate the action run table and stages panel