Improve UX for adding relationship form element
Clicking edit on a newly added (not saved to the db) field, then
clicking cancel will no longer delete the element. That will only happen
if you click cancel immediately after adding the new element.

The save button in the form element panel will always be enabled when
first configuring a new field input element. Previously the only way
to add a new relationship field without making an unnecessary config
edit was to click the x, since the Save button was disabled and the
cancel button would remove the element from the form.